Add loop to find a document from a list to support multiple recipes
