+++ Makefile
@@ -9,9 +9,10 @@ export GOARCH?=$(shell go env GOARCH)
 
 build:
 	@mkdir -p $(BUILD_ARCH)/$(BIN_DIR)
-	go build -o $(BUILD_ARCH)/$(BIN_DIR)/dp-recipe-api cmd/dp-recipe-api/main.go
+	go build -o $(BUILD_ARCH)/$(BIN_DIR)/dp-recipe-api main.go
+
 debug:
-	HUMAN_LOG=1 go run cmd/dp-recipe-api/main.go
+	HUMAN_LOG=1 go run main.go
 
 test:
 	go test -cover $(shell go list ./... | grep -v /vendor/)
